Resolved conflict in common/acdefault/default.go the contract abi and
bytecode had been updated on develop and conflicted with my removal.

The solidity version in Autonity.sol had been updated, so I needed to
update the solc version used in the Makefile.

The github actions had been modified, so I had to update them to ensure
that they all call 'make embed-autonity-contract'.

+++ consensus/tendermint/committee/committee_mock.go
@@ -136,10 +136,10 @@ func (mr *MockSetMockRecorder) Copy() *gomock.Call {
 }
 
 // F mocks base method
-func (m *MockSet) F() int {
+func (m *MockSet) F() uint64 {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "F")
-	ret0, _ := ret[0].(int)
+	ret0, _ := ret[0].(uint64)
 	return ret0
 }
 
@@ -150,10 +150,10 @@ func (mr *MockSetMockRecorder) F() *gomock.Call {
 }
 
 // Quorum mocks base method
-func (m *MockSet) Quorum() int {
+func (m *MockSet) Quorum() uint64 {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Quorum")
-	ret0, _ := ret[0].(int)
+	ret0, _ := ret[0].(uint64)
 	return ret0
 }
